refactor: eventTimer 분리
