+++ src/basic/main.basic.js
@@ -25,6 +25,7 @@ import {
   POINT_RATES,
   TIMER_INTERVALS,
 } from './constants';
+import { lightningSaleTimer, suggestSaleTimer } from './utils/eventTimers';
 
 let bonusPts = 0;
 let stockInfo;
@@ -84,50 +85,8 @@ function main() {
   onUpdateSelectOptions();
   handleCalculateCartStuff();
 
-  // 번개세일 팝업
-  const lightningDelay = Math.random() * TIMER_INTERVALS.LIGHTNING_DELAY;
-  setTimeout(() => {
-    setInterval(function () {
-      const luckyIdx = Math.floor(Math.random() * PRODUCT_LIST.length);
-      const luckyItem = PRODUCT_LIST[luckyIdx];
-      if (luckyItem.q > 0 && !luckyItem.onSale) {
-        luckyItem.val = Math.round((luckyItem.originalVal * 80) / 100);
-        luckyItem.onSale = true;
-        alert('⚡번개세일! ' + luckyItem.name + '이(가) 20% 할인 중입니다!');
-        onUpdateSelectOptions();
-        doUpdatePricesInCart();
-      }
-    }, TIMER_INTERVALS.LIGHTNING_SALE);
-  }, lightningDelay);
-
-  // 추천 팝업
-  const suggestDelay = Math.random() * TIMER_INTERVALS.SUGGEST_DELAY;
-  setTimeout(function () {
-    setInterval(function () {
-      if (cartDisp.children.length === 0) {
-      }
-      if (lastSel) {
-        let suggest = null;
-        for (let k = 0; k < PRODUCT_LIST.length; k++) {
-          if (PRODUCT_LIST[k].id !== lastSel) {
-            if (PRODUCT_LIST[k].q > 0) {
-              if (!PRODUCT_LIST[k].suggestSale) {
-                suggest = PRODUCT_LIST[k];
-                break;
-              }
-            }
-          }
-        }
-        if (suggest) {
-          alert('💝 ' + suggest.name + '은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!');
-          suggest.val = Math.round((suggest.val * (100 - 5)) / 100);
-          suggest.suggestSale = true;
-          onUpdateSelectOptions();
-          doUpdatePricesInCart();
-        }
-      }
-    }, TIMER_INTERVALS.SUGGEST_SALE);
-  }, suggestDelay);
+  lightningSaleTimer(onUpdateSelectOptions, doUpdatePricesInCart);
+  suggestSaleTimer(onUpdateSelectOptions, doUpdatePricesInCart, cartDisp, lastSel);
 }
 
 let sum;
